Moved viewer out of oceaninsight.py
